requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
aiohttp
aiofiles
pydantic
jinja2
termcolor
pytube
python-magic
beautifulsoup4
lxml
feedparser
requests
google-api-python-client
langchain
langchain-community
youtube-transcript-api
python-docx
python-docx-replace
isodate
PyMuPDF
Pillow
PyPDF2
reportlab
pytesseract 